feat : ShootStatusService 분리
